+++ src/Array.ts
@@ -683,3 +683,24 @@ export const reduceWhile = <A>(p: Predicate<A>) => <B>(
 
   return go(ys)(x)
 }
+
+/**
+ * Like ordinary array reduction, however this also takes a predicate that is
+ * evaluated before each step. If the predicate doesn't hold, the reduction
+ * short-circuits and returns the current accumulator value.
+ *
+ * @example
+ * import { reduceRightWhile } from 'fp-ts-std/Array';
+ * import { add } from 'fp-ts-std/Number';
+ * import { Predicate } from 'fp-ts/function';
+ *
+ * const isEven: Predicate<number> = n => n % 2 === 0;
+ * const reduceRightUntilOdd = reduceRightWhile(isEven);
+ *
+ * assert.strictEqual(reduceRightUntilOdd(add)(0)([2, 4, 7, 8, 10]), 18);
+ *
+ * @since 0.8.0
+ */
+export const reduceRightWhile = <A>(p: Predicate<A>) => <B>(
+  f: (x: A) => (y: B) => B,
+) => (x: B): ((ys: Array<A>) => B) => flow(A.reverse, reduceWhile(p)(f)(x))