+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to semantic versioning.
 
 ## 0.8.0 (_Unreleased_)
 
-- Add `reduceWhile` function to the `Array` module.
+- Add `reduceWhile` and `reduceRightWhile` functions to the `Array` module.
 
 ## 0.7.0 (2020-11-29)